pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ypbrief"
version = "0.1.0"
description = "Self-hosted scheduled YouTube podcast brief runner"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32",
    "click>=8.1",
    "PyYAML>=6.0",
    "fastapi>=0.110",
    "uvicorn>=0.30",
    "APScheduler>=3.10",
]

[project.optional-dependencies]
transcripts = ["yt-dlp>=2025.6.30", "webvtt-py>=0.5.1"]
llm = [
    "openai>=1.0",
    "anthropic>=0.49",
    "google-genai>=1.0",
]
dev = ["pytest>=8"]

[project.scripts]
ypbrief = "ypbrief.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
